+++ packages/expo/src/WebBrowser/ExponentWebBrowser.web.ts
@@ -1,0 +1,5 @@
+export default {
+  get name() {
+    return 'ExponentWebBrowser';
+  },
+};
